services/pool-generator/requirements.txt
celery[redis]==5.4.*
redis[hiredis]
python-dotenv
google-genai
groq
